Added a new 'spede-term' command to easily open serial terminals
To support connecting to serial terminals on the SPEDE target, a new
utility was needed to detect the psuedo terminal created for each COM
port. Currently, this uses minicom but potentially could use programs
in the future instead if desired.

Additionally, disabled the parallel port in spede-target by default
but gave an option to re-enable it.

Lastly, renamed the default and config arrays for multiple programs to
use the same names to prepare for combining common code into a single
library.